+++ adminforth/spa/src/components/Toast.vue
@@ -1,7 +1,8 @@
 <template>
     
 
-<div id="toast-default" class="flex  items-center w-full p-4 text-gray-500 rounded-lg shadow dark:text-gray-400 dark:bg-gray-800" role="alert"
+<div class="flex items-center w-full p-4 text-gray-500 rounded-lg shadow-lg dark:text-gray-400 dark:bg-gray-800 bg-white"
+    role="alert"
     :class="
         {
             'danger': 'bg-red-100',