+++ adminforth/spa/src/stores/filters.ts
@@ -3,7 +3,14 @@ import { defineStore } from 'pinia';
 
 export const useFiltersStore = defineStore('filters', () => {
     const filters: Ref<any[]> = ref([]);
-    
+    const sort: Ref<any> = ref({});
+
+    const setSort = (s: any) => {
+        sort.value = s;
+    }
+    const getSort = () => {
+        return sort.value;
+    }
     const setFilter = (filter: any) => {
         filters.value.push(filter);
     }
@@ -16,5 +23,5 @@ export const useFiltersStore = defineStore('filters', () => {
     const clearFilters = () => {
         filters.value = [];
     }
-    return {setFilter, getFilters, clearFilters, filters, setFilters}
+    return {setFilter, getFilters, clearFilters, filters, setFilters, setSort, getSort}
 })